added the ability to use left control to disable auto zoom out and made the vertex shader use floats for translate x y and z

+++ src/main/java/app/cleancode/graphics/Node.java
@@ -12,9 +12,9 @@ public Node() {
 protected abstract void glDraw();
 
 public final void draw() {
-	ShaderLoader.setShaderUniform("translateX", Float.valueOf(translateX).intValue());
-	ShaderLoader.setShaderUniform("translateY", Float.valueOf(translateY).intValue());
-	ShaderLoader.setShaderUniform("translateZ", Float.valueOf(translateZ).intValue());
+	ShaderLoader.setShaderUniform("translateX", translateX);
+	ShaderLoader.setShaderUniform("translateY", translateY);
+	ShaderLoader.setShaderUniform("translateZ", translateZ);
 	glDraw();
 }
